Add files via upload
Outline for Map and Place implementation. I included an additional level of abstraction and modified the previous code for Maps so that the changes are reflected. I also added the outlines for the new location tiles in the game. With his, movement should still be intact (minus teleportation), and the markets should be functional as I had them implemented before (definitely need to be looked at)

Also I was very tired so the variable declarations switch between public and private a lot since typing up a million getters/setters was getting annoying. Will fix after I sleep but they should all be private for best practice